+++ src/JK.Common/Extensions/DateTimeExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Globalization;
 using JK.Common.TypeHelpers;
 
 namespace JK.Common.Extensions
@@ -7,103 +6,53 @@ namespace JK.Common.Extensions
     /// <summary>Extension methods for the DateTime object.</summary>
     public static class DateTimeExtensions
     {
-        /// <summary>
-        /// Adds given number of business days to a date.
-        /// </summary>
+        /// <summary>Adds given number of business days to a date.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="days">Number of days to add (can be negative).</param>
         /// <returns>The date the given amount of business days from the start date.</returns>
-        public static DateTime AddWorkDays(this DateTime dt, int days) => new DateTimeHelper().AddWorkDays(dt, days);
+        public static DateTime AddWorkDays(this DateTime dt, int days) => DateHelper.AddWorkDays(dt, days);
 
-        /// <summary>
-        /// Calculates age of an individual.
-        /// </summary>
+        /// <summary>Calculates age of an individual.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="birthday">Date of birth.</param>
         /// <returns>Age from birth date to given date.</returns>
-        public static int CalculateAge(this DateTime dt, DateTime birthday) => new DateTimeHelper().CalculateAge(dt, birthday);
+        public static int CalculateAge(this DateTime dt, DateTime birthday) => DateHelper.CalculateAge(dt, birthday);
 
-        /// <summary>
-        /// Gets the abbreviated day name.
-        /// </summary>
+        /// <summary>Gets the abbreviated day name.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>Abbreviated day name.</returns>
-        public static string GetAbbreviatedDayName(this DateTime dt) => new DateTimeHelper().GetAbbreviatedDayName(dt);
+        public static string GetAbbreviatedDayName(this DateTime dt) => DateHelper.GetAbbreviatedDayName(dt);
 
-        /// <summary>
-        /// Gets the abbreviated day name.
-        /// </summary>
-        /// <param name="dt">Current DateTime object from extension method.</param>
-        /// <param name="currentCulture">Culture to use when getting day name.</param>
-        /// <returns>Abbreviated day name.</returns>
-        public static string GetAbbreviatedDayName(this DateTime dt, CultureInfo currentCulture) => new DateTimeHelper().GetAbbreviatedDayName(dt, currentCulture);
-
-        /// <summary>
-        /// Gets the abbreviated month name.
-        /// </summary>
+        /// <summary>Gets the abbreviated month name.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>Abbreviated month name.</returns>
-        public static string GetAbbreviatedMonthName(this DateTime dt) => new DateTimeHelper().GetAbbreviatedMonthName(dt);
+        public static string GetAbbreviatedMonthName(this DateTime dt) => DateHelper.GetAbbreviatedMonthName(dt);
 
-        /// <summary>
-        /// Gets the abbreviated month name.
-        /// </summary>
-        /// <param name="dt">Current DateTime object from extension method.</param>
-        /// <param name="currentCulture">Culture to use when getting month name.</param>
-        /// <returns>Abbreviated month name.</returns>
-        public static string GetAbbreviatedMonthName(this DateTime dt, CultureInfo currentCulture) => new DateTimeHelper().GetAbbreviatedMonthName(dt, currentCulture);
-
-        /// <summary>
-        /// Gets the full name of a given day.
-        /// </summary>
+        /// <summary>Gets the full name of a given day.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>Full day name.</returns>
-        public static string GetDayName(this DateTime dt) => new DateTimeHelper().GetDayName(dt);
+        public static string GetDayName(this DateTime dt) => DateHelper.GetDayName(dt);
 
-        /// <summary>
-        /// Gets the full name of a given day.
-        /// </summary>
-        /// <param name="dt">Current DateTime object from extension method.</param>
-        /// <param name="currentCulture">Culture to use when getting day name.</param>
-        /// <returns>Full day name.</returns>
-        public static string GetDayName(this DateTime dt, CultureInfo currentCulture) => new DateTimeHelper().GetDayName(dt, currentCulture);
-
-        /// <summary>
-        /// Gets the full name of a month.
-        /// </summary>
+        /// <summary>Gets the full name of a month.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>Full month name.</returns>
-        public static string GetMonthName(this DateTime dt) => new DateTimeHelper().GetMonthName(dt);
+        public static string GetMonthName(this DateTime dt) => DateHelper.GetMonthName(dt);
 
-        /// <summary>
-        /// Gets the full name of a month.
-        /// </summary>
-        /// <param name="dt">Current DateTime object from extension method.</param>
-        /// <param name="currentCulture">Culture to use when getting month name.</param>
-        /// <returns>Full month name.</returns>
-        public static string GetMonthName(this DateTime dt, CultureInfo currentCulture) => new DateTimeHelper().GetMonthName(dt, currentCulture);
-
-        /// <summary>
-        /// Determines whether or not a given date is between (inclusive) the given start and end dates.
-        /// </summary>
+        /// <summary>Determines whether or not a given date is between (inclusive) the given start and end dates.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <param name="start">Start of date range to check</param>
         /// <param name="end">End of date range to check</param>
         /// <returns>True if date falls within range, otherwise false</returns>
-        public static bool IsBetween(this DateTime dt, DateTime start, DateTime end) => new DateTimeHelper().IsBetween(dt, start, end);
+        public static bool IsBetween(this DateTime dt, DateTime start, DateTime end) => DateHelper.IsBetween(dt, start, end);
 
-        /// <summary>
-        /// Determines if given date is a weekday.
-        /// </summary>
+        /// <summary>Determines if given date is a weekday.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>True if is a weekday, otherwise false.</returns>
-        public static bool IsWeekday(this DateTime dt) => new DateTimeHelper().IsWeekday(dt);
+        public static bool IsWeekday(this DateTime dt) => DateHelper.IsWeekday(dt);
 
-        /// <summary>
-        /// Determines if given date is a weekday.
-        /// </summary>
+        /// <summary>Determines if given date is a weekday.</summary>
         /// <param name="dt">Current DateTime object from extension method.</param>
         /// <returns>True if is a weekend, otherwise false.</returns>
-        public static bool IsWeekend(this DateTime dt) => new DateTimeHelper().IsWeekend(dt);
+        public static bool IsWeekend(this DateTime dt) => DateHelper.IsWeekend(dt);
     }
 }